Fixes problem with “second_pass” function
In the course of the aggregation step where communitites are merged into "super nodes" multiple edges between nodes have to be accounted for.
This is crucial as it effects properties of the graph like the degree of the nodes, which in turn the modularity computation relies on and can consequently lead to wrong results.
This is solved through updating the weight attribute of the graph giving it the correct degree.